requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
Django==5.1.3
django-ipware==7.0.1
idna==3.10
pycparser==2.22
PyNaCl==1.5.0
python-decouple==3.8
python-ipware==3.0.0
requests==2.32.3
sqlparse==0.5.2
urllib3==2.2.3
